Adhering constants pi and e the new format
