+++ ivy/functional/backends/jax/array_api/constants.py
@@ -1,0 +1,6 @@
+#global
+import jax.numpy as jnp
+
+
+e = jnp.e
+pi = jnp.pi
