+++ ivy/functional/backends/mxnet/array_api/constants.py
@@ -1,0 +1,5 @@
+#global
+import math 
+
+e = math.e
+pi = math.pi
