+++ ivy/functional/backends/mxnet/core/math.py
@@ -6,8 +6,6 @@ Collection of MXNet math functions, wrapped to fit Ivy syntax and signature.
 import mxnet as _mx
 import math as _math
 
-e = _math.e
-pi = _math.pi
 sin = lambda x: _math.sin(x) if isinstance(x, float) else _mx.nd.sin(x)
 cos = lambda x: _math.cos(x) if isinstance(x, float) else _mx.nd.cos(x)
 tan = lambda x: _math.tan(x) if isinstance(x, float) else _mx.nd.tan(x)
