+++ ivy/functional/backends/numpy/array_api/constants.py
@@ -1,0 +1,5 @@
+#global
+import numpy as np
+
+e = np.e
+pi = np.pi
