+++ ivy/functional/backends/numpy/core/math.py
@@ -9,8 +9,6 @@ try:
 except (ImportError, ModuleNotFoundError):
     _erf = None
 
-e = _np.e
-pi = _np.pi 
 sin = _np.sin
 cos = _np.cos
 tan = _np.tan
