+++ ivy/functional/backends/tensorflow/array_api/constants.py
@@ -1,0 +1,5 @@
+# global
+import tensorflow as tf
+
+e = tf.experimental.numpy.e
+pi = tf.experimental.numpy.pi
