+++ ivy/functional/backends/torch/array_api/constants.py
@@ -1,0 +1,5 @@
+# global
+import torch
+
+e = torch.e
+pi = torch.pi
