+++ ivy/functional/backends/torch/core/math.py
@@ -6,8 +6,6 @@ Collection of PyTorch math functions, wrapped to fit Ivy syntax and signature.
 import math
 import torch as _torch
 
-e = _torch.e
-pi = _torch.pi 
 
 def sin(x):
     if isinstance(x, float):
